move interfaces back into original files
