+++ src/Page.ts
@@ -39,13 +39,81 @@ import {
 import { Accessibility } from './Accessibility';
 import { TimeoutSettings } from './TimeoutSettings';
 import { FileChooser } from './FileChooser';
-import { ConsoleMessage } from './ConsoleMessage'
-import {
-  ScreenshotOptions, PaperFormat, Metrics, WaitForOptions, MediaFeature,
-  ScreenshotClip, PDFOptions
-} from './types'
+import { ConsoleMessage } from './ConsoleMessage';
+import { PuppeteerLifeCycleEvent } from './LifecycleWatcher';
 
 const writeFileAsync = helper.promisify(fs.writeFile);
+
+interface Metrics {
+  Timestamp?: number;
+  Documents?: number;
+  Frames?: number;
+  JSEventListeners?: number;
+  Nodes?: number;
+  LayoutCount?: number;
+  RecalcStyleCount?: number;
+  LayoutDuration?: number;
+  RecalcStyleDuration?: number;
+  ScriptDuration?: number;
+  TaskDuration?: number;
+  JSHeapUsedSize?: number;
+  JSHeapTotalSize?: number;
+}
+
+interface WaitForOptions {
+  timeout?: number;
+  waitUntil?: PuppeteerLifeCycleEvent | PuppeteerLifeCycleEvent[];
+}
+
+interface MediaFeature {
+  name: string;
+  value: string;
+}
+
+interface ScreenshotClip {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+}
+
+interface ScreenshotOptions {
+  type?: 'png' | 'jpeg';
+  path?: string;
+  fullPage?: boolean;
+  clip?: ScreenshotClip;
+  quality?: number;
+  omitBackground?: boolean;
+  encoding?: string;
+}
+
+interface PDFMargin {
+  top?: string | number;
+  bottom?: string | number;
+  left?: string | number;
+  right?: string | number;
+}
+
+interface PDFOptions {
+  scale?: number;
+  displayHeaderFooter?: boolean;
+  headerTemplate?: string;
+  footerTemplate?: string;
+  printBackground?: boolean;
+  landscape?: boolean;
+  pageRanges?: string;
+  format?: string;
+  width?: string | number;
+  height?: string | number;
+  preferCSSPageSize?: boolean;
+  margin?: PDFMargin;
+  path?: string;
+}
+
+interface PaperFormat {
+  width: number;
+  height: number;
+}
 
 const paperFormats: Record<string, PaperFormat> = {
   letter: { width: 8.5, height: 11 },
